// src/lib/A-Workflows/A-Workflows.types.ts
export type A_TYPES__WorkflowValueSource =
    | { value: any; from?: never; fn?: never; args?: never }
    | { from: string; value?: never; fn?: never; args?: never }
    | {
        fn: A_WorkflowValueFn | `${A_WorkflowValueFn}`;
        args?: Array<A_TYPES__WorkflowValueSource>;
        value?: never;
        from?: never;
    };
// ...
export type A_TYPES__WorkflowRemap = Record<string, A_TYPES__WorkflowValueSource>;
// ...
export type A_TYPES__WorkflowCondition = {
    fn: A_WorkflowConditionFn | `${A_WorkflowConditionFn}`;
    left?: A_TYPES__WorkflowValueSource;
    right?: A_TYPES__WorkflowValueSource;
    conditions?: Array<A_TYPES__WorkflowCondition>;
};